NYC vehicle crash analysis
NYC vehicle accident data is studied in this proposal. The dataset analyzed consists of a merge of three different csv files, containing information about the crash, people, and vehicle involved in accidents. The project idea consists in identifying the regions in NYC which are the most dangerous as far as accidents are concerned. Final findings would show the most critical NYC neighborhoods for each road user type (driver, pedestrian, cyclist, etc). This can be achieved by combining the ZIP code of the accident with NYC neighborhood dataset.
After identifying the most critical regions, a deeper step has to be taken to identify the reasons why these places are so dangerous. Is it related to the time of the day? Is there a major highway nearby? Is it due to rush hour? Is this a bohemian area? Is it due to road conditions? With the root cause identified, solutions will be proposed that can be used by NYC to reduce the number of accidents, culminating in the reduction of governments costs (estimation of $35 billion annually only in the US. Source: Costs of Crashes to Government, United States, 2008)
This initial analysis code checks the ratio of accidents that resulted in injury or deaths over the total number of accidents. The data is then split into the boroughs, and accident percentage calculated over the years available in the dataset. Manhattan is found to be the borough with the least amount of fatal accidents over the years.